task 0

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,21 +1,22 @@
 #include <stdio.h>
-#include <search_algos.h>
+#include "search_algos.h"
 /**
  * linear_search - goes through the whole array
- * @array: pointer to an array 
+ * @array: pointer to an array
  * @size: size of an array
  * @value: the value searched for
+ * Return: a value on Successs, -1 if failed
  */
 int linear_search(int *array, size_t size, int value)
 {
-int i = 0;
+size_t i = 0;
 if (array == NULL)
 {
 return (-1);
 }
-while (i < int(size))
+while (i < size)
 {
-printf("Value checked array[%d] = [%d]\n", i, array[i]);
+printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 if (array[i] == value)
 {
 return (i);
